use original watchOS view
